Skills/spatial-worlds/assets-pipeline/requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
requests>=2.31.0
